pyproject.toml
[project]
name = "wetter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "sqlalchemy>=2.0.44",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "bs4>=0.0.2",
    "pymysql>=1.0.0",
    "dotenv>=0.9.9",
    "plotly>=6.5.0",
    "dash>=2.0.0",
    "mysql-connector-python>=9.5.0",
]
